normals.cpp:191:30: fix warning: comparison of integer expressions of different signedness

+++ bundled_deps/admesh/admesh/normals.cpp
@@ -188,7 +188,7 @@ void stl_fix_normal_directions(stl_file *stl)
     	// Get next facet to fix from top of list.
     	if (head->next != tail) {
       		facet_num = head->next->facet_num;
-            assert(facet_num < stl->stats.number_of_facets);
+            assert(static_cast<uint32_t>(facet_num) < stl->stats.number_of_facets);
       		if (norm_sw[facet_num] != 1) { // If facet is in list mutiple times
         		norm_sw[facet_num] = 1; // Record this one as being fixed.
         		++ checked;
